requirements.txt
asgiref==3.2.10
attrs==20.3.0
backcall==0.2.0
coverage==5.5
decorator==5.0.6
Django==3.1.2
djangorestframework==3.12.4
Faker==8.1.0
iniconfig==1.1.1
Markdown==3.3.4
mixer==7.1.2
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
pytest-django==4.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
